requirements.txt
streamlit==1.33.0
spacy==3.7.2
sentence-transformers==2.6.1
scikit-learn==1.3.2
pdfplumber==0.10.3
python-docx==1.1.0
nltk==3.8.1
torch>=1.13.0
beautifulsoup4==4.12.3
requests==2.31.0